refactor buildModal
