+++ packages/page-spy-plugin-data-harbor/src/assets/locale.ts
@@ -14,10 +14,13 @@ const source = {
     invalid: '无法上传',
     copied: '已复制调试连接',
     selectPeriod: '选择时间段',
+    periodDuration: '每段时长',
     remark: '备注',
     remarkPlaceholder: '上传后，可以看到备注信息',
-    periodFrom: '从',
-    periodTo: '到',
+    from: '从',
+    to: '到',
+    refreshed: '已刷新',
+    minutes: '分钟',
   },
   en: {
     title: 'Offline log',
@@ -30,12 +33,15 @@ const source = {
     success: 'Succeed',
     fail: 'Failed',
     invalid: 'Cannot upload',
-    copied: 'Copied',
+    copied: 'Replay url copied',
     selectPeriod: 'Select period',
+    periodDuration: 'Each duration',
     remark: 'Remark',
     remarkPlaceholder: 'After uploading, you can view the remark information.',
-    periodFrom: 'From',
-    periodTo: 'To',
+    from: 'From',
+    to: 'To',
+    refreshed: 'Refreshed',
+    minutes: 'minutes',
   },
 };
