+++ packages/page-spy-plugin-data-harbor/src/assets/svg.ts
@@ -29,3 +29,9 @@ export const downloadPeriodsSvg = `
     <path fill="currentColor" d="M10.975 21.95q-1.9-.2-3.537-1.062t-2.838-2.2t-1.887-3.063t-.688-3.65q0-3.875 2.563-6.712T11 2v2q-3.025.425-5 2.688t-1.975 5.287T6 17.263t4.975 2.687zm1-4.95L6.95 11.95l1.425-1.425l2.6 2.6V7h2v6.125l2.575-2.575L16.975 12zm1 4.95v-2q1.075-.15 2.063-.575t1.837-1.075l1.45 1.45q-1.175.925-2.525 1.488t-2.825.712m3.95-16.3Q16.05 5 15.063 4.575T13 4V2q1.475.15 2.825.713t2.5 1.487zm2.8 12.65l-1.4-1.425q.65-.85 1.05-1.837t.55-2.063h2.05q-.2 1.475-.75 2.838t-1.5 2.487m.2-7.325q-.15-1.075-.55-2.062t-1.05-1.838l1.4-1.425q.95 1.125 1.525 2.488t.725 2.837z"/>
   </svg>
 `;
+
+export const refreshSvg = `
+  <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 24 24">
+    <path fill="currentColor" d="M3.5 12q0-3.616 2.664-6.058T12.5 3.5v-.692q0-.242.217-.354t.43.03l2.163 1.62q.323.242.323.646t-.323.646l-2.164 1.62q-.212.142-.429.03t-.217-.354V6Q9.86 6 7.93 7.718T6 12q0 .883.247 1.713t.73 1.556q.217.323.167.69t-.348.583l-.407.298q-.373.293-.818.227t-.711-.444q-.668-1.017-1.014-2.2Q3.5 13.241 3.5 12m8 8.5v.692q0 .243-.217.354q-.217.112-.43-.03l-2.162-1.62q-.324-.242-.324-.646t.323-.646l2.164-1.62q.212-.142.429-.03t.217.354V18q2.64 0 4.57-1.718T18 12q0-.883-.247-1.722q-.247-.84-.73-1.566q-.217-.323-.167-.69t.348-.583l.407-.298q.373-.293.818-.225t.712.442q.661 1.036 1.01 2.21T20.5 12q0 3.616-2.664 6.058T11.5 20.5"/>
+  </svg>
+`;
