+++ packages/page-spy-plugin-data-harbor/src/harbor/base.ts
@@ -1,4 +1,5 @@
 import type { SpyMessage } from '@huolala-tech/page-spy-types';
+import { PeriodItem } from './blob';
 
 export type DataType =
   | 'console'
@@ -7,7 +8,21 @@ export type DataType =
   | 'storage'
   | 'rrweb-event';
 
-export type Actions = 'download' | 'upload';
+export type Actions =
+  | 'upload'
+  | 'download'
+  | 'upload-periods'
+  | 'download-periods';
+
+export interface WholeActionParams {
+  clearCache?: boolean;
+  remark?: string;
+}
+export interface PeriodActionParams {
+  from: PeriodItem;
+  to: PeriodItem;
+  remark?: string;
+}
 
 export type CacheMessageItem = Pick<
   SpyMessage.MessageItem<SpyMessage.DataType, any>,
